--- client/src/components/Navbar/Original-Navbar.jsx
@@ -1,12 +1,0 @@
-// import './Style-Navbar.css';
-
-// export const Navbar = () => {
-//   return (
-//     <nav>
-//       <div>Logos</div>
-//       <div>Searchbar</div>
-//       <div>Post</div>
-//       <div>Sign In</div>
-//     </nav>
-//   );
-// };
